src/annotate/requirements.txt
Django==1.8.4
GDAL==1.10.0
Pillow==2.9.0
appnope==0.1.0
argparse==1.3.0
cffi==1.2.1
cryptography==1.0
decorator==4.0.2
django-extensions==1.5.5
django-jsonfield==0.9.15
enum34==1.0.4
gnureadline==6.3.3
idna==2.0
ipaddress==1.0.14
ipython==4.0.0
ipython-genutils==0.1.0
ndg-httpsclient==0.4.0
path.py==7.6.1
pexpect==3.3
pickleshare==0.5
pyOpenSSL==0.15.1
pyasn1==0.1.8
pycparser==2.14
requests==2.7.0
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
wsgiref==0.1.2
